German dict creates File if file not found
If the file that the German dictionary is opened on is not found, unwort
used to just fail and say that the file was not found. But if you want
to open a new dictionary you'd have to create the file manually and
leave it empty, which doesn't make sense, so this creates a new file for
you if it is not found.